backend/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
pydantic==2.12.5
google-genai==1.55.0
